+++ pkg/webui/console/store/actions/identity-server.js
@@ -1,4 +1,4 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2020 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,18 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import snapshotDiff from 'snapshot-diff'
+import { createRequestActions } from './lib'
 
-global.snapshotDiff = snapshotDiff
-
-/* eslint-disable no-console */
-const originalConsoleError = console.error
-console.error = function(message, ...args) {
-  console.log(message)
-  if (/(Invalid prop|Failed prop type|Failed context type)/gi.test(message)) {
-    throw new Error(message)
-  }
-
-  originalConsoleError.apply(console, [message, ...args])
-}
-/* eslint-enable no-console */
+export const GET_IS_CONFIGURATION_BASE = 'GET_IS_CONFIGURATION'
+export const [
+  {
+    request: GET_IS_CONFIGURATION,
+    success: GET_IS_CONFIGURATION_SUCCESS,
+    failure: GET_IS_CONFIGURATION_FAILURE,
+  },
+  {
+    request: getIsConfiguration,
+    success: getIsConfigurationSuccess,
+    failure: getIsConfigurationFailure,
+  },
+] = createRequestActions(GET_IS_CONFIGURATION_BASE)