+++ pkg/webui/console/views/admin-user-management-add/index.js
@@ -22,6 +22,8 @@ import PageTitle from '@ttn-lw/components/page-title'
 import { withBreadcrumb } from '@ttn-lw/components/breadcrumbs/context'
 import Breadcrumb from '@ttn-lw/components/breadcrumbs/breadcrumb'
 
+import withRequest from '@ttn-lw/lib/components/with-request'
+
 import UserDataForm from '@console/components/user-data-form'
 
 import sharedMessages from '@ttn-lw/lib/shared-messages'
@@ -29,14 +31,25 @@ import PropTypes from '@ttn-lw/lib/prop-types'
 
 import { createUser } from '@console/store/actions/users'
 import { attachPromise } from '@console/store/actions/lib'
+import { getIsConfiguration } from '@console/store/actions/identity-server'
+
+import {
+  selectPasswordRequirements,
+  selectIsConfigurationFetching,
+} from '@console/store/selectors/identity-server'
 
 @connect(
-  undefined,
+  state => ({
+    passwordRequirements: selectPasswordRequirements(state),
+    fetching: selectIsConfigurationFetching(state),
+  }),
   {
     createUser: attachPromise(createUser),
     navigateToList: () => push(`/admin/user-management/`),
+    getConfiguration: () => getIsConfiguration(),
   },
 )
+@withRequest(({ getConfiguration }) => getConfiguration(), ({ fetching }) => fetching)
 @withBreadcrumb('admin.user-management.add', () => {
   return <Breadcrumb path={`/admin/user-management/add`} content={sharedMessages.add} />
 })
@@ -44,6 +57,11 @@ export default class UserManagementAdd extends Component {
   static propTypes = {
     createUser: PropTypes.func.isRequired,
     navigateToList: PropTypes.func.isRequired,
+    passwordRequirements: PropTypes.passwordRequirements,
+  }
+
+  static defaultProps = {
+    passwordRequirements: undefined,
   }
 
   @bind
@@ -61,12 +79,15 @@ export default class UserManagementAdd extends Component {
   }
 
   render() {
+    const { passwordRequirements } = this.props
+
     return (
       <Container>
         <PageTitle title={sharedMessages.userAdd} />
         <Row>
           <Col lg={8} md={12}>
             <UserDataForm
+              passwordRequirements={passwordRequirements}
               onSubmit={this.onSubmit}
               onSubmitSuccess={this.onSubmitSuccess}
               onSubmitFailure={this.onSubmitFailure}