+++ sdk/js/src/api/http.js
@@ -13,8 +13,9 @@
 // limitations under the License.
 
 import axios from 'axios'
+import { cloneDeep, get } from 'lodash'
 
-import { URI_PREFIX_STACK_COMPONENT_MAP } from '../util/constants'
+import { URI_PREFIX_STACK_COMPONENT_MAP, STACK_COMPONENTS_MAP } from '../util/constants'
 import EventHandler from '../util/events'
 
 import stream from './stream/stream-node'
@@ -112,7 +113,17 @@ class Http {
       return response
     } catch (err) {
       if ('response' in err && err.response && 'data' in err.response) {
-        throw err.response.data
+        const error = cloneDeep(err.response.data)
+        // Augment the default error with config entries as well as the stack component
+        // abbreviation that threw an error.
+        // TODO: Consider changing this, see https://github.com/TheThingsNetwork/lorawan-stack/issues/3424.
+        error.request_details = {
+          url: get(err, 'response.config.url'),
+          method: get(err, 'response.config.method'),
+          stack_component: parsedComponent,
+        }
+
+        throw error
       } else {
         throw err
       }
@@ -141,14 +152,14 @@ class Http {
    * Extracts The Things Stack component abbreviation from the endpoint.
    *
    * @param {string} endpoint - The endpoint got for a request method.
-   * @returns {string} One of {is|as|gs|js|ns}.
+   * @returns {string} The stack component abbreviation.
    */
   _parseStackComponent(endpoint) {
     try {
       const component = endpoint.split('/')[1]
       return Boolean(URI_PREFIX_STACK_COMPONENT_MAP[component])
         ? URI_PREFIX_STACK_COMPONENT_MAP[component]
-        : 'is'
+        : STACK_COMPONENTS_MAP.is
     } catch (err) {
       throw new Error('Unable to extract The Things Stack component:', endpoint)
     }